Some fixes

+++ sdk/src/client/api/block_builder/input_selection/mod.rs
@@ -270,8 +270,8 @@ impl InputSelection {
         mut inputs: Vec<InputSigningData>,
         outputs: &[Output],
         slot_index: SlotIndex,
-        min_committable_age: u64,
-        max_committable_age: u64,
+        min_committable_age: u32,
+        max_committable_age: u32,
     ) -> Result<Vec<InputSigningData>, Error> {
         // initially sort by output to make it deterministic
         // TODO: rethink this, we only need it deterministic for tests, for the protocol it doesn't matter, also there
