+++ sdk/src/types/block/slot/commitment_id.rs
@@ -8,7 +8,7 @@ impl_id!(pub SlotCommitmentId, 40, "Identifier of a slot commitment.");
 impl SlotCommitmentId {
     pub fn index(&self) -> SlotIndex {
         // PANIC: taking the last 8 bytes of 40 bytes is safe.
-        u64::from_le_bytes(
+        u32::from_le_bytes(
             self.0[Self::LENGTH - core::mem::size_of::<SlotIndex>()..]
                 .try_into()
                 .unwrap(),
